+++ sekoia_valhalla_integration_modules/models.py
@@ -5,6 +5,5 @@ DEMO_API_KEY = "1" * 64
 
 class SekoiaValhallaIntegrationModuleConfiguration(BaseModel):
     api_key: str = Field(default=DEMO_API_KEY)
-    base_url: str = Field(default="https://valhalla.nextron-systems.com")
     sekoia_api_key: str = Field(default="")
     sekoia_base_url: str = Field(default="https://api.sekoia.io")